requirements.txt
django==4.2
djangorestframework==3.14.0
django-query-counter==0.2.2
drf-standardized-errors==0.12.5
drf-yasg==1.21.5
python-dotenv==1.0.0
pandas==2.0.1
